core(test) cra -> vite 테스트
