+++ package.json
@@ -4,22 +4,22 @@
   "description": "AI-based source code analysis and documentation tool.",
   "author": "Your Name",
   "private": true,
-  "main": "build/main.js",
+  "main": "build/electron/main.js",
   "homepage": "./",
   "dependencies": {
     "@testing-library/jest-dom": "^6.6.4",
-    "@testing-library/react": "^16.3.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "electron-is-dev": "^2.0.0",
-    "glob": "9.3.5",
+    "glob": "^10.4.2",
     "jszip": "^3.10.1",
-    "react": "^19.1.1",
-    "react-dom": "^19.1.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.6.1",
     "reactflow": "^11.11.4",
     "recharts": "^3.1.2",
-    "typescript": "^4.9.5",
+    "typescript": "4.9.5",
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
@@ -28,22 +28,27 @@
     "@types/jest": "^27.5.2",
     "@types/jszip": "^3.4.0",
     "@types/node": "^16.18.126",
-    "@types/react": "^19.1.9",
-    "@types/react-dom": "^19.1.7",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.0",
     "@types/react-syntax-highlighter": "^15.5.13",
     "concurrently": "^8.2.2",
     "electron": "^31.0.1",
     "electron-builder": "^24.13.3",
+    "nodemon": "^3.1.10",
+    "rimraf": "^6.0.1",
+    "ts-node": "^10.9.2",
     "wait-on": "^7.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
-    "postbuild": "if not exist build\\main.js copy public\\main.js build\\main.js",
+    "build:react": "react-scripts build",
+    "build:electron": "tsc --project tsconfig.electron.json",
+    "clean": "rimraf build",
+    "build": "npm run clean && npm run build:react && npm run build:electron",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "electron:start": "concurrently \"npm start\" \"wait-on http://localhost:3000 && electron public/main.js\"",
-    "electron:build": "npm run build && npm run postbuild && electron-builder"
+    "electron:start": "concurrently \"npm start\" \"wait-on http://localhost:3000 && nodemon\"",
+    "electron:build": "npm run build && electron-builder"
   },
   "optionalDependencies": {
     "fsevents": "*"
@@ -53,8 +58,7 @@
     "productName": "소스추출기",
     "asar": true,
     "files": [
-      "build/**/*",
-      "public/preload.js"
+      "build/**/*"
     ],
     "directories": {
       "output": "dist",
@@ -67,7 +71,7 @@
       "target": "dmg"
     },
     "extraMetadata": {
-      "main": "build/main.js"
+      "main": "build/electron/main.js"
     }
   },
   "browserslist": {
